requirements.txt
Django==5.0.7
djangorestframework==3.15.2
drf_yasg==1.21.7
Faker==26.0.0
python-dotenv==1.0.1
psycopg2-binary==2.9.9